Search IMPROVED
Indicative(temporary) but not intrusive(permanent) search highlight is
enabled.

Motion commands `n` and `N` become consistent regardless of the previous
search direction, which possibly violates a strong convention in
exchange for a more comfortable habit. But a change for the similar
commands `fFtT` is under consideration yet.

Incremental search is improved by adding a search specific command line
interface, which allows e.g. jumping to the next match during a search
before it's considered complete.